+++ CHANGELOG.md
@@ -5,6 +5,12 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 - [DemoApp/SDK] #222 Update the Button page to list the available button types
 - [SDK] Bug #96 Add Small Card 
+- [DemoApp] #228 Change Chip to switch on List screen for "SecondLineOfText" and "LeadingImage" options 
+- [SDK/DemoApp] #237 Define a Grid of small cards module and use it in application for components
+- [SDK/DemoApp] #182 Add card title first
+- [SDK/DemoApp] #237 Define a Grid of small cards module and use it in application for components 
+- [DemoApp] #235 Rename Module called Card View to Card Collections (2)
+- [SDK] Bug #96 Add Small Card
 
 ## [0.5.0] - 2022-09-02